+++ src/routes/_examples/AreaStacked.svelte
@@ -1,6 +1,6 @@
 <script>
-	import { LayerCake, Svg, flatten } from 'layercake';
-	import { stack } from 'd3-shape';
+	import { LayerCake, Svg, flatten, stack } from 'layercake';
+
 	import { scaleOrdinal } from 'd3-scale';
 	import { format, precisionFixed } from 'd3-format';
 	import { timeParse, timeFormat } from 'd3-time-format';
@@ -15,29 +15,28 @@
 	const xKey = 'month';
 	const yKey = [0, 1];
 	const zKey = 'key';
-
-	const parseDate = timeParse('%Y-%m-%d');
+	const xKeyCast = timeParse('%Y-%m-%d');
 
 	const seriesNames = Object.keys(data[0]).filter(d => d !== xKey);
 	const seriesColors = ['#ff00cc', '#ff7ac7', '#ffb3c0', '#ffe4b8'];
 
+	const formatTickX = timeFormat('%b. %-d')
+	const formatTickY = d => format(`.${precisionFixed(d)}s`)(d);
+
+	/* --------------------------------------------
+	 * Cast data
+	 */
 	data.forEach(d => {
-		d[xKey] = typeof d[xKey] === 'string' ? parseDate(d[xKey]) : d[xKey];
+		d[xKey] = typeof d[xKey] === 'string'
+			? xKeyCast(d[xKey])
+			: d[xKey];
+
 		seriesNames.forEach(name => {
 			d[name] = +d[name];
 		});
 	});
 
-	/* --------------------------------------------
-	 * Create a stacked data structure
-	 */
-	const stackData = stack()
-		.keys(seriesNames);
-
-	const series = stackData(data);
-
-	const formatTickX = timeFormat('%b. %-d')
-	const formatTickY = d => format(`.${precisionFixed(d)}s`)(d);
+	const stackedData = stack(data, seriesNames);
 </script>
 
 <style>
@@ -62,8 +61,8 @@
 		zScale={scaleOrdinal()}
 		zDomain={seriesNames}
 		zRange={seriesColors}
-		flatData={flatten(series)}
-		data={series}
+		flatData={flatten(stackedData)}
+		data={stackedData}
 	>
 		<Svg>
 			<AxisX