+++ src/routes/_examples/ColumnStacked.svelte
@@ -1,6 +1,6 @@
 <script>
-	import { LayerCake, Svg, flatten, uniques } from 'layercake';
-	import { stack } from 'd3-shape';
+	import { LayerCake, Svg, flatten, uniques, stack } from 'layercake';
+
 	import { scaleBand, scaleOrdinal } from 'd3-scale';
 	import { format, precisionFixed } from 'd3-format';
 
@@ -18,18 +18,18 @@
 	const seriesNames = Object.keys(data[0]).filter(d => d !== xKey);
 	const seriesColors = ['#00e047', '#7ceb68', '#b7f486', '#ecfda5'];
 
+	/* --------------------------------------------
+	 * Cast data
+	 */
 	data.forEach(d => {
 		seriesNames.forEach(name => {
 			d[name] = +d[name];
 		});
 	});
 
-	const stackData = stack()
-		.keys(seriesNames);
-
-	const series = stackData(data);
-
 	const formatTickY = d => format(`.${precisionFixed(d)}s`)(d);
+
+	const stackedData = stack(data, seriesNames);
 </script>
 
 <style>
@@ -47,17 +47,17 @@
 
 <div class="chart-container">
 	<LayerCake
- 			padding={{ top: 0, right: 0, bottom: 20, left: 20 }}
- 			x={d => d.data[xKey]}
- 			y={yKey}
- 			z={zKey}
- 			xScale={scaleBand().paddingInner([0.02]).round(true)}
- 			xDomain={uniques(data, xKey)}
+			padding={{ top: 0, right: 0, bottom: 20, left: 20 }}
+			x={d => d.data[xKey]}
+			y={yKey}
+			z={zKey}
+			xScale={scaleBand().paddingInner([0.02]).round(true)}
+			xDomain={uniques(data, xKey)}
 			zScale={scaleOrdinal()}
 			zDomain={seriesNames}
 			zRange={seriesColors}
- 			flatData={flatten(series)}
- 			data={series}
+			flatData={flatten(stackedData)}
+			data={stackedData}
 	>
 		<Svg>
 			<AxisX
@@ -71,5 +71,4 @@
 			<ColumnStacked/>
 		</Svg>
 	</LayerCake>
-
 </div>